Minimize raw example
Still need to connect the OLED to a Trinket and actually test it. Also still need to figure out why it's only writing 32 bytes.

When compiling for the Trinket it automatically uses the TinyWireM library instead of Wire. Here is the compiler output when compiling for the Micro and Trinket

Sketch uses 5382 bytes (18%) of program storage space. Maximum is 28672 bytes.
Global variables use 391 bytes (15%) of dynamic memory, leaving 2169 bytes for local variables. Maximum is 2560 bytes.

Sketch uses 860 bytes (16%) of program storage space. Maximum is 5310 bytes.
Global variables use 93 bytes of dynamic memory.