First button test
I found a small pushbutton and put it on a mini breadboard. The adjacent pins (on the same side of the gap) are disconnected until the button is pressed. I connected one pin to ground and the other to pin 4 (pins 2 and 3 were being used for I2C for the SSD1306 OLED display) on my Arduino Micro (on the other mini breadboard). I set pin 4 to INPUT_PULLUP so it will read HIGH until the button is pressed. Pressing the button will connect that pin to ground. This path will have less resistance (no 20-50kΩ resistor). To begin I'll just have the built-in LED turn on when the button is pressed. Remember, the pin will read HIGH when the button is NOT pressed, and LOW when it IS pressed. I could have used the define LED_BUILTIN instead of the numeric literal 13.