Test using interrupts for button
I moved the jumper wire from pin 4 to pin 7, as the Arduino Micro supports interrupts on pins 0, 1, 2, 3, and 7. Pins 2 and 3 are being used for I2C. My method is named isr for "interrupt service routine" (ISR is reserved for other use). Integer n must be declared volatile as its state will change outside of the primary routine. I've attached the interrupt to the FALLING edge because the INPUT_PULLUP pin mode means it will be HIGH until the button is pressed, FALLING to LOW.